Add @Values for ports via application.props
