+++ hw16-Frontend/src/main/resources/application.properties
@@ -1,1 +1,4 @@
+messageServer.host=localhost
+messageServer.port=8081
+
