+++ hw16-Messageserver/src/main/resources/application.properties
@@ -1,1 +1,4 @@
 spring.main.web-application-type=NONE
+
+messageServer.port=8081
+
